refactor(query): extract range constraint analysis
